refactor: move payment related features into public-form/payment
