+++ frontend/src/features/public-form/components/FormPaymentPage/components/CreatePaymentIntentFailureBlock.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useMemo, useRef } from 'react'
 import { Box, Flex, Stack, Text, VisuallyHidden } from '@chakra-ui/react'
 
-import { usePublicFormContext } from '../../public-form/PublicFormContext'
+import { usePublicFormContext } from '~features/public-form/PublicFormContext'
+
 import { FormPaymentPageProps } from '../FormPaymentPage'
 
 // Make sure to call `loadStripe` outside of a component’s render to avoid
