+++ frontend/src/features/public-form/components/FormPaymentPage/components/PaymentSuccessSvgr.tsx
