+++ frontend/src/features/public-form/components/FormPaymentPage/stripe/StripeReceiptContainer.tsx
