+++ frontend/src/features/public-form/components/FormPaymentPage/stripe/components/StripeDownloadReceiptBlock.tsx
