+++ frontend/src/features/public-form/components/FormPaymentPage/stripe/utils.ts
